# openapi.yaml
openapi: 3.0.0
info:
  title: Task Manager API
  description: RESTful API for managing tasks (create, read, update, delete).
  version: 1.0.0

servers:
  - url: http://localhost:3000

paths:
  /api/tasks:
    get:
      summary: Get all tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

    post:
      summary: Add a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTask'
      responses:
        '200':
          description: Task added confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /api/tasks/{id}:
    put:
      summary: Update a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTask'
      responses:
        '200':
          description: Task updated confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

    delete:
      summary: Delete a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task deleted confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string

    NewTask:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
        description:
          type: string

    UpdateTask:
      type: object
      required:
        - title
        - description
        - status
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string